GitHub OAuth2 implementation as Gerrit HTTP Auth.
First integration point for allowing  Gerrit to authenticate users through GitHub OAuth.
It relies on the installation of the github-oauth Java library under the $GERRIT_SITE/lib
in order filter all the HTTP requests through the GitHub OAuth 2.0 secure authentication.

In order to configure the GitHub authentication you need to:
- copy the github-oauth into $GERRIT_SITE/lib
- copy the github-plugin into $GERRIT_SITE/plugins
- run the Gerrit init steps by choosing HTTP Auth
- set the GitHub parameters during the GitHub custom init step

Change-Id: I0898994169fec0b477dc8944045ceba9526a4d9c