Moving dependencies to child module.
Instead of getting all dependencies at parent Maven module
they are now defined at child level.
This allows to only include the ones effectively needed
and avoid duplications when using shaded jars.

Adding as well a copy of the Gerrit InitStep interface
as it is missing in the Gerrit API package.

Change-Id: I8c8edbc05599d1e507d269baf5d95f32b99870e7