[GERRITHUB-4] GitHub repositories selection and clone.
Select the GitHub repositories to be automatically
configured for replication and review with Gerrit.
When accepted, clone all remote repositories
using JGit to the corresponding Gerrit projects
as organization/repository.

Page for selection and repository clone is all
made with jQuery and AJAX calls to allow maximum
responsiveness of the page and avoid time-outs
due to long-running clone operations.

In case of failures, page is idempotent and can be
used again for cloning the remaining repositories.
All incomplete clones are automatically cleaned up
so that can be retried at a later time.

Change-Id: Iba6338bb5fbe813d78c87c444815d94814144228