[GERRITHUB-4] Set Gerrit refs/meta/config on imported repos.
Create initial Project properties (project.config
and groups) on refs/meta/config whenever a GitHub
project gets imported.

In order to allow a non-Gerrit administrator to have
full control over the imported project, assign
extra owner and GitHub-like permissions to the individual
user  that performed the operation.

P.S. In order to allow Gerrit to assign permissions
to an individual user, a SingleUserGroup plugin is
needed to be installed.

Change-Id: I596b2e80b4d9519668a1ab289d6c950139d6a922