[GERRITHUB-4] Import/replicate from organisations.
Shows a single-select drop-down on the GitHub
repositories import page with the list of organisations
that current users belongs to.
Once the organisation is selected, the list of
GitHub repositories available for import and replication
is updated, allowing to select and trigger the import.

By selecting different organisations, it is possible
to perform multiple GitHub repositories imports sessions.

Change-Id: Ib425c5012fb661d807b3c7545d117b00f8829163