[RFC] Pull-request importer screen.
GitHub pull request importer screen: repos 
configured in Gerrit for replication with GitHub 
are scanned and list of opened GitHub pull requests 
are displayed for selection.

Pull requests previously imported in Gerrit are 
omitted from the list. Pull requests with new commits
are displayed, as the new commits will be imported
as new dependent changes.

Change-Id: I7e6a02859e6d5a7e76691b52625cdcacd2da2960