Redirect rule for GitHub import wizard.
Introduce the ability to control forwarding or
redirecting to the next wizard step.

Forward rule:
wizardFlow = account.gh => repositories.html

Redirect rule example:
wizardFlow = account.gh R> /

The value of forward is the ability
to keep the HTTP Request with all associated parameters
and send them to the target page / servlet.
The drawback is the Browser URL not being updated
and the target page / servlet should behave
consistently with the different target URL in the request
and associated parameters.

Change-Id: I37bd5724183ac0042c1cc59cd0e97f930f94388e