Auto-update secure.config OAuth token when needed.
GitHub OAuth can decide to expire an existing token
and generate a new one, based on time and requested
scope change: this may impact existing credentials
stored in secure.config and then block replication
to GitHub repos.

With this enhancement the secure.config is checked
at login and updated if needed.

NOTE: we could potentially have concurrency issues
(even if the update is a synchronised method) because
of replication auto-config with concurrent OAuth token 
updates. FileBasedConfig already takes care of File-level
locking on save.

Change-Id: I41fe6eb8020c2fdecad90232ddd27af5f8a995d7