Use new Gerrit 2.9-SNAPSHOT patch-set inserter
The new Gerrit master patch-set inserter relies
on the ChangeControl to get the current user for 
being used as patch-set submitter.

This is typically correct but in case of GitHub
the patch is created "on behalf" of the pull request
on GitHub: we need to then simulate this delegation
of responsibilities as we want Gerrit to notify the
original Pull Request user and NOT the "robot" used
for creating the patches in Gerrit.

Change-Id: Ibbafeb680e007d635944dba48924fbb840dbfb4a