Adding Google storage repository for Maven dependencies
People are reporting that Gerrit API dependencies cannot be found
without having to build Gerrit itself.
This change allows them to rely on GoogleStorage repository
and avoid to build Gerrit up-front.

Additionally the direct fetching of the Gerrit 2.8 released API
highlighted a problem in the Change constructor that has now been
fixed.

Change-Id: I8fec5c4e220b85ae4703db1fc70fd66abc2d57eb