Avoid use of @SessionScoped for GitHub login
Guice's session scoped beans rely on the fact
that GuiceServlet has captured the request and 
handled in the current thread.
This assumption is not true when a request gets
filtered outside GuiceServlet, as it happens for
the GitHub OAuth Filter.

With this change the GitHub login is directly
obtained from the HTTP Servlet Request without
relying on GuiceServlet on the active request
thread.

Change-Id: Id9d687e9cb0cbab413e71ab5c369e64a10ec74f6