Share GitHub login on Gerrit filter and plugin.
Allows to sign-in to GitHub only once during
Gerrit login and then share the GitHub OAuth token
within the plugin screens.
Previously the two logins were completely isolated
and the user had to connect twice.

With this change is not possible anymore to have
multiple logins with different scopes: not a regression
but a simplification IMHO, as having multiple sessions
with multiple permissions have lead to confusion
on why GitHub permissions are requested twice.

Change-Id: I8aeb131704ee015d67b56670295c231dd6db1478