ScopedProvider is now more similar to standard Guice's Provider
By hiding the HttpServletRequest inside the HttpSessionProvider,
the ScopedProvider<T> can be now 100% compatible with the standard
Guice's Provider<T> interface.

Moving as well from constructor injection to field injection so
that the concrete implementation do not have to replicate the
constructor injection definition. Code is more compact and extensible.

The ability to get an instance by passing an actual HttpServletRequest
is kept in order to allow the standard servlet filter (outside the Guice
domain) to retrieve the object in the current session, without
having to request a Provider<HttpServletRequest> which would fail.

Change-Id: I20f32fc42ca05e9a4a140b64a1d5693d83ecf125