Align GitHub OAuth token expire with GerritAccount
Calculate the GerritOAuth Cookie.maxAge in the same
way that Gerrit Cached WebSessions expire.
This allows to avoid the unpleasant situations where
Gerrit session is active but all RESTFul API calls
are failing because of the missing OAuth cookie.

This was perceived as a "Internal Error" pop-up
on the Gerrit UX when the OAuth token expired.

Change-Id: I0d1f5d3be1f9257ee3b8ee5f0ac1dd809f1727fd