Avoid OAuth with REST and RPC calls and static resources
When Gerrit UX is making REST API calls to the backend
by providing both an GerritAccount cookie and
a X-GERRIT-AUTH HTTP Header, the call can be safely
propagate to Gerrit for validation without having to 
necessarily re-create a new GerritOAuth cookie.

Gerrit UX is unable anyway to complete the GitHub
OAuth handshake, there is no value then in getting
the calls redirected to GitHub as it would just break
the user interface.

Change-Id: Ie1566364be6eb647609971b88e2a30c78e012afe