Refactor OAuth login handshake: remove usage of GerritOAuth cookie
The GerritOAuth cookie is now obsolete as we do rely on a shared
GHLogin object in the HTTP Session. Whenever the session expires or
is lost, the situation is detected and automatically recovered by
triggering a new OAuth handshake with GitHub.

The session GHLogin object is used as well to remember the user
scopes and username for HTTP Header enrichment.

Change-Id: Ibede5c16995f6659c8401c6d5083d44f1185e0ba