FIX IndexOutOfBound in redirection / forwarding
When redirecting to the next step in the GitHub
import wizard an IndexOutOfBound exception was 
thrown when using relative URL in gerrit.config
for the wizard steps.

Example:
[github]
wizardFlow = repositories-next.gh => pullrequests.html

This was due to the inconsistence of requestURI 
vs. contextPath. Target URI needs always to be 
absolute and have contextPath as prefix of requestURI. 

Change-Id: Ib5ef32378f9b351bf568aaf615b67cf864904a32