Native ChangeControl for user instead of delegate
We can now get a ChangeControl object on behalf
of another user, so that we can create Changes
owned by the Pull Request sender instead of the
currently logged in Gerrit user.

There is no need of using a ChangeControl delegate
with package access modifier in order to impersonate
another user.

Change-Id: Icdeca9875b9724c49b34980e3c4224f22b11a5b8