RFC: Consistent indentation in HTML templates
Having consistent indentation makes the HTML templates easier to
read, and thus easier to spot problems.

- Replace tabs and change indentation level to 2 spaces. Having a mix
of tabs and space makes it difficult to notice when the indentation
is wrong.

- Make indentation consistent with HTML tag nesting. Indenting in this
way makes it easier to notice when a closing tag is missing.

- Remove unnecessary line breaks.

- Add missing comments.

Only the account page is modified.  If this kind of change is
acceptable, I will do the others too.

Change-Id: I0eaddc22d0ca086edc6f89f261b5e078b334af11