Revert "Set plugin name to ‘github’" This reverts commit ef88bd14d2b2ff852af2d938ccb077998f7a9685.
It seems that GitHub Plugin makes assumptions on the plugin
name in the manifest to be matching exactly the plugin 
file name.

This is BAD and the assumption needs to be removed, but
in the meantime we need to keep that assumption valid 
and avoiding the plugin name override in the manifest.

Change-Id: Ie8c1e197ae8e55c1e092ab0e044d095830500748