Update to master GitHub API for fixing for user’s private repos list
GitHub API contains a bug when listing user’s private repos: when
list is paginated the listing includes ONLY the user’s public ones.
When list is NOT paginated all the repos are returned instead.

The lucamilanesio’s fork of GitHub API on master (1.57-SNAPSHOT)
contains the fix for this problem. Hopefully will be soon merged
and released by Kohsuke into a 1.57 final.

Change-Id: I3089c61ba5a930425c8162a58711e0b9700ef4e7