Default “Everyone” Team for public-only OAuth scopes
When GitHub users do not allow to share their private information
the organisation’s teams are not visible and thus no groups are
available in Gerrit.

In order to at least allow to use the organisation membership itself
as group a new default “Everyone” Team is added which represents
all the people belonging to the GitHub organisation.

In this way at least a ON/OFF permissions scheme is possible
allowing extra permissions to everyone belonging to the public
organisation.

Change-Id: I7545948a2f4343d9cc6a9299b14576640223b59f