Fix OAuth callback URL for non-root Gerrit canonical URL
When Gerrit canonical URL is not root (e.g. myserver.com/gerrit)
we need to append the oauth suffix to it in order to be
correctly routed and managed through the reverse proxy.

Additionally the method is never called using a relative URL
and used only to initialise the config fields during object
construction.
It is better to remove the “else” branch, make the method static 
and private to prevent misuse from other parts of the code.

Change-Id: I00b9be65eedd42779a0b70bb200323912ca620d7