Correct management of OAuth token serialisation returned by GitHub
When the OAuth/Phase-2 fails (last POST to GitHub API with the
one-time-code returned to the /oauth callback) we need to
read the error codes and descriptions returned and avoid
to perform the subsequent login operations.

Additionally the OAuth token bean specs should respect the common
camelCase notation adopted in Gerrit.

Change-Id: Ib7e42aba761a6f4ff4f7870509113b8d53efa764