Filter out RO GitHub repositories for replication
When setting up Gerrit->GitHub replication using the repositories
wizard, it does not make sense to list the ones that are read-only
and then risking to amend replication.config with settings that
would never work.

It is much better to make sure that what the repos we list are the
ones that actually we are authorized to replicate to.

Change-Id: I976cfd02ea45032bd9236dc73cac5f62996baca4