Replication target by username not on organization
Previously the replication targets were defined on a per-organization
basis, assuming that only an elected user would be defined for running
the replication process.

However things are more complex and sometimes organizations are not
centrally managed and, especially with OpenSource, different people may
operate independently and have different permissions to different repos.

It is then quite difficult (if not impossible) to define a single user
to carry out replication for all repos. A safer approach is to define
a one-target-per-user logic where each individual can potentially carry
out replication on only a sub-set of repos inside an organisation.

This avoids the potential risk of overwriting the replication user
inside an organisation and allow different users to replicate their own
repos.

Change-Id: I52d4ea9883335a35a15f32a080b1d0b9796d79c6