Update patch set 1
Patch Set 1:

> @David: merge was already done. Are pushing this only to get the
 > merge tracked in Git?

Exactly.  As long as git doesn't see a merge commit it will try to merge everything again, causing conflicts.

This is merged with '-s ours' so after this the stable-2.11 branch is the same as before.  The only difference is that now any further commits on stable-2.10 can be merged easier.

I know you said you're focussing on stable-2.11 now, but you never know what might come up.  I think it's cleaner to make this merge.

Patch-set: 1