Save the last replication error to a project's JSON
Whenever a replication error is detected through Gerrit stream events
it is saved under the plugin’s data directory with the full JSON
of the failed replication event on the associated ref.
Multiple errors on different refs will be stored onto different
JSON files on different refs directories.

When replication resumes and a new successful replication event
for that ref will be received, the JSON gets removed.

This simple JSON file can be used for reporting problems directly
to the project owner via e-mail or crontab or a dedicated GUI.

Change-Id: Ie9eeea82de669134bdfe71e409a8c90f2d3d89d3