Adapt repositories listing to latest GitHub API
A couple of days ago GitHub started to return *ALL* the repos
a user collaborates to when requesting /user/repos whilst previously
returned only the ones owned by current user.

We need to use the v1.70 of Kohsuke's GitHub API in order to filter
out all the repos but the ones really owned by current user.

Change-Id: I9ab56968c44a633efb01834b495444e493638172