Adapt repositories listing to latest GitHub API
A couple of days ago GitHub started to return *ALL* the repos
a user collaborates to when requesting /user/repos whilst previously
returned only the ones owned by current user.

We need to use an enhanced version of Kohsuke's GitHub API
(with PR#203 [1]) in order to filter out all the repos but the ones
really owned by current user.

NOTE: This fix requires an ad-hoc build of Kohsuke's GitHub API
      with [1] cherry-picked on it, until it will be merged and released

[1] https://github.com/kohsuke/github-api/pull/203

Change-Id: I9ab56968c44a633efb01834b495444e493638172