[WIP] Add webhook to automatically import pull requests
Adds a new servlet which handles webnook events sent from Github.
The servlet forwards the payload to an EventHandler based on the type of
the event, then the handler does the actual work.

Currently "ping" and "pull_request" events are supported

TODO(yugui) Extract hard-coded configuration to GithubConfig
TODO(yugui) Move the import operation into a background thread?
TODO(yugui) Write a journal record before importing so that Gerrit does
not lose the import operation even if it is shutdown during the import

Change-Id: I389f78c702dee1c59192768acffcff49e0c80a62