JVM-wide proxy settings for GitHub API http calls
When using Gerrit behind a proxy, the http.proxyUrl setting defines 
the proxy details. However the GitHub API do not use that setting
and instead rely on the JVM-wide system properties.

We do need to get the Gerrit settings and set them at JVM level
in order for the GitHub API to work properly.

Change-Id: Ifd5bc8c4614e32cfaed43665fb5bdcff92085ac6