Allows user changing scope within a Gerrit session
Even after the user has been already authenticated with GitHub with a
default OAuth scope, it is possible now to visualize the current scope
and changing it on-the-fly.

This is particularly useful during repositories import where a user
realises that the initial scope wasn’t enough to display private
organizations or repositories.

Change-Id: I5af497571db7482bd74ddf25ffccc26fea3bcaa3