Allow plugin reload at runtime
When Gerrit needs to load a plugin, copies the jar file to a temporary
location so that overwrites do not impact the normal class and 
resource loading. This makes the plugin reloadable at runtime.

Until now the GitHub plugin has used instead the original jar for
loading its Velocity resources which made the plugin NOT reloadable.

This change makes the detection of the Jar file where the plugin is 
loaded from discoverable at runtime and the magic happens: GitHub plugin
becomes fully reloadable allowing hot-deploys.

Change-Id: Icd16059de2522d1fcc5c5e6f4e15c06b3d38d699