Fix broken build on Gerrit master (2.13-SNAPSHOT)
Since the merge of I2d961f8d it is not possible to find patch-sets
by SHA1 anymore. We need to leverage the Lucene index to check
if a pull-request commit has been already imported or not.

MenuItem is now imported from com.google.gerrit.extensions.client to
be able to compile against the latest Gerrit master.

Change-Id: I9868b3a0c1c58bb2ea027be58a561252773277fa