AccountCache by name isn't reliable: get from id
It seems that on current Gerrit master, the account cache loader
isn't actually loading the data anymore but just relies on the Lucene
indexes to retrieve the account info.

It seems that some of the information is missing from there: the custom
GitHub OAuth external ID can't be found and needs to be retrieved
using the cache by id.

This is a temporary workaround: the main issue is to fix the Gerrit's
loading cache to load the real data and not just the index.

Additionally, even the cache by id sometimes is stale.
Before failing, it is worth trying again by evicting the accounts cache.

Change-Id: I943cfeac0f8a1c717a4f37e67d29ea83f8fa059a