Fix exception during Gerrit init on authType
When initializing Gerrit, do not fail when processing
the selection of the authType: even if only HTTP and OAUTH are
valid with the GitHub plugin, it is incorrect to fail with
an IllegalArgumentException.

Use instead the EnumSet of the AuthType allowed values
to specify the only valid values with GitHub authentication.

Change-Id: I3abcd821cfa33550c1d5102f2f91cfa25fe3f2d1