Fix default scopes resolution
The introduction of the scopes description in Change-Id: I80f30fcb6
failed to update the key of the default scopes, causing the plugin
to malfunction when no specific scopes configuration were defined.

Fix the problem by replacing the "scopes" string with a proper
ScopeKey associated with the expected key created when defining
a default section for scopes in gerrit.config.

Change-Id: I4cff4b93d25a77c50fb0cc93f9234a40a9f5f8d5