Bump Apache Velocity to 2.3
Apache Velocity 2.3 is the latest version
and relies on commons lang v3 as Gerrit 3.6
requires.

Also use the new artifact name velocity-engine-core and
use the Gerrit's static assets servlet for serving binary
files as the new Velocity resource manager in v2 isn't able
to read binary content.

Inject the Plugin instance for allowing the access and load
of the plugin's own resources.

Bug: Issue 15897
Change-Id: I5802c16167d9a0b8591cd42e0a7e9ef429e4d5a3