Encrypt oauth token at rest
The oauth token provided by Github is used together with the
`github_oauth` schema, to generate a new external identity for the
Gerrit user.

The external-id is then persisted on disk as a note under the
`refs/meta/external-ids` ref of the `All-Users.git` repo, as documented
in [1].

Obtaining access to the `All-Users.git` repository constitutes a
security risk, as the token could be used to impersonate the user
associated to it.

This security risk is described in detail [2].

Avoid using plain oauth tokens for storing github external-ids by
encrypting them first.

Bug: Issue 16192

[1] https://osm.etsi.org/gerrit/Documentation/config-accounts.html#external-ids
[2] https://github.blog/2022-04-15-security-alert-stolen-oauth-user-tokens/

Change-Id: I69c2add9170159ab8b9a31098cf35f184e678401