Generate default key when no key is configured
When no single key is configured password (16 bytes long) is generated
and stored in '$site/data/@PLUGIN@/default.key' and returned 'current'
key's passwordDevice. It is used from now on until new key is
configured.

Note that 'IllegalStateException' is being thrown when 'passwordDevice'
is not configured for given key configuration.

Bug: Issue 16192
Change-Id: I99e4024ecf19d828deb2d84b40fe1dd7e1abe23d