Set parent project at creation time for imported repos from GitHub
The configuration of the Gerrit parent project should be done
right at creation time for avoiding to have cached ACLs on the
previous project hierarchy.

Gerrit does not evict the whole chain of project hierarchy
for avoiding expensive cache coherence evaluation. However, the
import of a project with the wrong public hierarchy
may lead the project to be public right after the import.

The cache eviction would then eventually make the repository
private, however the exposure to external access for a few minutes
represent a security risk and therefore must be avoided at all
costs.

Move the parent project definition at the GitCloneStep stage
so that the right hierarchy is in place at the time of the
git fetch from GitHub.

Bug: Issue 16727
Change-Id: Id1a6c4809eed43ba618ed13e2c3161a8968dd063