Fix cache eviction upon project import from GitHub
The import phase from GitHub has three steps:
1. create a Gerrit project with the correct public/private parent
2. configure the project's ACLs
3. import existing commits from GitHub

The project cache would not needed to be evicted, in theory.
However, after step 1 the project as not ACLs and its entry is
cache in the projects cache. The  step 2 therefore would need to
wait the projects cache TTL for allowing the ACLs to be taken
into consideration.

Add an explicit evict after step 2, so that ACLs have immediate
effect.

Change-Id: I02880564562f170ef0a73bcec56dd37d39930d7a