Allow login in a different domain
A new login flow has been provided to allow requests from the domain i.e
subdomain-2.domain-1.io to be redirected to a different domain i.e
subdomain-1.domain-1.io in order to achieve the login. Once the the login
has been done successfuly, the flow will be redirected to the initial
domain subdomain-2.domain-1.io.

TODO: explain the motivation of this change and add documentation in
README.md

Change-Id: I8e037f3c5580be2a63f517df769878e8b6e81921