Build OAuth redirect URL when X-Forwarded-Host is present
In the past, when a Gerrit multi-site topology was used featuring a
primary domain, i.e example.com (serving as a geo-location load
balancer), along with multiple Gerrit sites like review-1.example.com
and review-2.example.com, the initiation of the GitHub sign-in flow
within any of the Gerrit sites triggered a redirection to GitHub,
referred to as the user's GitHub identity request [1]. During this
process, the redirect_uri query parameter was constructed using
information from the gerrit.canonicalWebUrl property defined in the
etc/gerrit.config file [2]. It's important to note that this property
occasionally contained a URL with a host matching the primary domain,
example.com. Consequently, when users attempted to sign in via GitHub,
they were redirected to this main domain.

However, with this updated approach, the redirect_uri is now constructed
based on the X-Forwarded-Host header if it is present in the request
(otherwise from gerrit.canonicalWebUrl). This means that when a sign-in
flow is initiated, for instance, from review-1.example.com, the
Forwarded-Host header will accurately contain the value
review-1.example.com. As a result, the redirection URL will utilize this
value as its host.

References:
[1] https://docs.github.com/en/apps/oauth-apps/building-oauth-apps/authorizing-oauth-apps#web-application-flow
[2] https://gerrit-review.googlesource.com/Documentation/config-gerrit.html

Bug: Issue 297231231
Change-Id: I233824a202041c7a7d1905ef784ae2b6f8d23160