Honour the cookieDomain also for the OAuth scope selection
Gerrit defines the cookieDomain for associating the authentication
cookie with a service; however, the cookie domain was not honoured
for the scope selection, making more difficult for users to switch
between different sites.

Example:
- User john logged in to site-1.gerrit.mycompany.com
- Gerrit cookie associated to .gerrit.mycompany.com
- User john login/logout to site-2.gerrit.mycompany.com
- The OAuth scope selection is lost

The OAuth scope cookie must respect the same cookieDomain defined
in the auth section of gerrit.config.

Change-Id: I0317bd38badb04c12a959973e80ffe726b3df168