Report authentication progress to the user
The GitHub OAuth API can get really slow, taking tens of seconds to
respond. Which is fine when users do see that there's something going
on in the backgournd.

Unfortunately current implementation of scope selection and sing-in
process gives no feedback about onging OAuth request. For some users
this may feel like something is broken or Gerrit site is slow.

To improve user experience we add progress "indicators" in both cases.

In the scope selection, the "Login >" button will be disabled after
clicking and its text will also change to 'Waiting for GitHub...'.

In case of the 'Sign-in' link in Gerrit UI, we use the new 'auth-link'
extension point to replace that link with custom component. The
'gr-github-oauth-progress' will wait 550ms and then show a loading
spinner (blocking the whole site) saying 'Waiting for GitHub...'.

In both cases users are now informated that the authentication is taking
place and they should wait for it to finish.

Bug: Issue 302826306
Change-Id: Ic4a8dc24d9133ac59dfe65858ac7ec8510cf4f09