Allow keeping custom configs in FanoutReplicationConfig
Some of the replication plugin default choices may not be suitable
for users of GitHub that need some specific replication settings.

Historically the default of te replication plugin were not suitable
for pushing to a non-Gerrit repository, hence the fragility of
generating and assuming defaults.

With this change, allow to keep any ad-hoc customisation on the
generated replication configuration files.

Change-Id: Ic451e934b1f8ef062c502eaad3c3f1bf63813556