Configure virutal domain specific scopes
When using both `github` and `virutalhost` plugins together is may be
required to have different scope selections for each site. Currently
the `github` plugin does not allow host specific scope configuration.

This patch adds the ability to use config subsection to configure
different scopes per site. For example:

[github "second.site.co"]
  scopeRedonly = USER_EMAIL
  scopeReadonlySequence = 0
  scopeReadonlyDescription = Share only email address

will configure only a single `USER_EMAIL` scope for `second.site.co`.

In order to make this work, we also change the redirection flow, as
currently `/login` will redirect to the main canonical web url. Which
means that we cannot discover which site is used and set scopes on scope
selection page. Now we'll keep user on the same domain name.

Bug: Issue 307761450
Change-Id: I24973cb2914052f23e684854211e1a21011dfecf