Refresh GitHub groups upon Gerrit successful login
GitHub groups are cached and persisted on disk, for preventing
GitHub outages to impact GerritHub.io. The persistent cache may
however cause security and inconsistency issues with the original
copy of the groups and memebership on GitHub.

Make sure that groups are refreshed every time a user is logging
on Gerrit Code Review successfully.

Because of the lack of login notification from Gerrit, the solution
needs to be based on the presence of the Set-Cookie header in
the HTTP response, hence the group refresh needs to use a
ServletFilter that is invoked for every request heading to Gerrit.

Bug: Issue 308979847
Change-Id: Ie5d012fb7c34e3edbb5576b53de32fd5b140ec17