Redirect to login on anonymous access to profile pages
When an unauthorized session tries to access any of the profile pages,
server will return status code 500 and a generic "Server Error" page
will be rendered. Instead, user should be redirected to the login page.

To fix this we simply check if user is authorized in
`VelocityViewServlet` and if not, redirect to the `auth.loginUrl`.

Bug: Issue 311627062
Change-Id: I3264f50500733e277ef10ca1a2ff36287f89fc99