Redirect user to previously visited page after login
When a user logs in from a change page, it is expected to be back on the
same page after the successful login. Gerrit's `/login` endpoint already
supports redirecting to the previously visited page, but that
information was not forwarded from the frontend.

To update the `href` in the login URL, we need to subscribe for
`nav-report` events. But as the event is emitted, before
`window.location` is updated, we also need to delay reading of the
current browser URL, until all other events are processed. This way, by
using `setTimeout`, we can reliably update the login link `href`.

But: Issue 325838103
Change-Id: Icc6498b48fac011a2bc2e92700ff216d781eb993